requirements.txt
# OCR System Requirements

# Core ML/Vision
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.13.0
Pillow>=9.5.0

# Evaluation & Metrics
jiwer>=3.0.1  # For CER/WER calculation
editdistance>=0.7.0

# Data & Processing
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM (optional, for post-processing)
openai>=0.27.0
requests>=2.31.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0

# Jupyter (optional, for exploration)
jupyter>=1.0.0
jupyterlab>=3.6.0
ipywidgets>=8.0.0

# Code Quality (optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
